Remove testnet::public_parameters duplication
