+++ kate/Cargo.toml
@@ -19,7 +19,6 @@ hex = { version = "0.4", default-features = false }
 kate-recovery = { path = "recovery", default-features = false, optional = true }
 log = { version = "0.4.8", optional = true }
 num_cpus = { version = "1.13.0", optional = true }
-once_cell = { version = "1.9.0", default-features = false }
 rand = { version = "0.8.4", default-features = false, optional = true }
 rand_chacha = { version = "0.3", default-features = false, optional = true }
 rand_core = { version = "0.5", default-features = false }
