+++ kate/recovery/Cargo.toml
@@ -11,6 +11,7 @@ derive_more = "0.99.17"
 dusk-bytes = "0.1.6"
 dusk-plonk = { git = "https://github.com/maticnetwork/plonk.git", tag = "v0.12.0-polygon-2" }
 getrandom = { version = "0.2", features = ["js"] }
+once_cell = { version = "1.9.0", default-features = false }
 rand = "0.8.4"
 rand_chacha = "0.3"
 serde = { version = "1.0", features = ["derive"] }
