+++ kate/src/lib.rs
@@ -125,30 +125,4 @@ impl sp_std::convert::TryInto<Dimensions> for BlockDimensions {
 	}
 }
 
-#[cfg(feature = "std")]
-pub mod testnet {
-	use std::{collections::HashMap, sync::Mutex};
-
-	use da_primitives::BlockLengthColumns;
-	use dusk_plonk::commitment_scheme::kzg10::PublicParameters;
-	use once_cell::sync::Lazy;
-	use rand::SeedableRng;
-	use rand_chacha::ChaChaRng;
-
-	static SRS_DATA: Lazy<Mutex<HashMap<usize, PublicParameters>>> =
-		Lazy::new(|| Mutex::new(HashMap::new()));
-
-	pub fn public_params(max_degree: BlockLengthColumns) -> PublicParameters {
-		let max_degree = max_degree.as_usize();
-		let mut srs_data_locked = SRS_DATA.lock().unwrap();
-		srs_data_locked
-			.entry(max_degree)
-			.or_insert_with(|| {
-				let mut rng = ChaChaRng::seed_from_u64(42);
-				PublicParameters::setup(max_degree, &mut rng).unwrap()
-			})
-			.clone()
-	}
-}
-
 // vim: set noet nowrap
